🔥 Remove: "불필요한 파일들 삭제"
Ref : #47
